html/interaction/focus/focusgroup/tentative/forward-navigation/moves-to-next-item-in-shadow-dom.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focusgroup - Focus moves to next item within Shadow DOM.</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com/">
<link rel="help" href="https://github.com/MicrosoftEdge/MSEdgeExplainers/blob/main/Focusgroup/explainer.md">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="../resources/focusgroup-utils.js"></script>

<div id=root focusgroup>
  <span id=item1 tabindex=0>item1</span>
  <div id=fg2 focusgroup=extend>
    <template shadowroot=open>
      <span id=item2 tabindex=-1>item2</span>
    </template>
  </div>
</div>

<script>

  promise_test(async t => {
    var item1 = document.getElementById("item1");
    var fg2 = document.getElementById("fg2");
    var shadowRoot = fg2.shadowRoot;
    var item2 = shadowRoot.getElementById("item2");

    await focusAndKeyPress(item1, kArrowDown);
    assert_equals(document.activeElement, fg2);
    assert_equals(shadowRoot.activeElement, item2);

    await focusAndKeyPress(item1, kArrowRight);
    assert_equals(document.activeElement, fg2);
    assert_equals(shadowRoot.activeElement, item2);
  }, "When the focus is set on a focusgroup item, an arrow key press should move the focus to the next item even if that next item is within a ShadowDOM.");

</script>
